# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TripDifficulty": {
				"enum": [
					"easy",
					"moderate",
					"challenging",
					"strenuous",
					"extreme"
				],
				"type": "string"
			},
			"TrekStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"Trek": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"coverMediaId": {
						"type": "string"
					},
					"mapMediaId": {
						"type": "string"
					},
					"galleryMediaIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"region": {
						"type": "string"
					},
					"durationDays": {
						"type": "number",
						"format": "double"
					},
					"maxAltitudeMeters": {
						"type": "number",
						"format": "double"
					},
					"difficulty": {
						"$ref": "#/components/schemas/TripDifficulty"
					},
					"priceUSD": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/TrekStatus"
					},
					"totalBookings": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"reviewsCount": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"galleryImages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"mapImage": {
						"type": "string"
					},
					"shortDesc": {
						"type": "string"
					},
					"bestSeason": {
						"type": "string"
					},
					"startEndLocation": {
						"type": "string"
					},
					"accommodation": {
						"type": "string"
					},
					"meals": {
						"type": "string"
					},
					"groupSizeRange": {
						"type": "string"
					},
					"permitsRequired": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"inclusionsText": {
						"type": "string"
					},
					"exclusionsText": {
						"type": "string"
					},
					"itinerary": {
						"items": {
							"$ref": "#/components/schemas/TripItineraryDay"
						},
						"type": "array"
					},
					"faqs": {
						"items": {
							"$ref": "#/components/schemas/TripFaq"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/TripReview"
						},
						"type": "array"
					},
					"addonsText": {
						"type": "string"
					},
					"usefulInfoText": {
						"type": "string"
					},
					"departureDates": {
						"items": {
							"$ref": "#/components/schemas/TripDepartureDate"
						},
						"type": "array"
					},
					"packageFiles": {
						"items": {
							"$ref": "#/components/schemas/TripPackageFile"
						},
						"type": "array"
					},
					"metaTitle": {
						"type": "string"
					},
					"metaDescription": {
						"type": "string"
					},
					"keywords": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"title",
					"slug",
					"categoryId",
					"region",
					"durationDays",
					"maxAltitudeMeters",
					"difficulty",
					"priceUSD",
					"status",
					"totalBookings",
					"rating",
					"reviewsCount",
					"shortDesc",
					"bestSeason",
					"startEndLocation",
					"accommodation",
					"meals",
					"groupSizeRange",
					"permitsRequired",
					"inclusionsText",
					"exclusionsText",
					"itinerary",
					"faqs",
					"reviews",
					"addonsText",
					"usefulInfoText",
					"departureDates",
					"packageFiles",
					"metaTitle",
					"metaDescription",
					"keywords"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginationMeta": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"nextPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"prevPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"lastPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count",
					"currentPage",
					"nextPage",
					"prevPage",
					"lastPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Trek-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Trek"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Trek"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_boolean_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"items": {
									"type": "boolean"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TourType": {
				"enum": [
					"cultural_heritage",
					"luxury_wellness",
					"wildlife_safari",
					"helicopter_tour",
					"day_tour",
					"other"
				],
				"type": "string"
			},
			"TourStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Tour-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Tour"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Tour"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ClimbingGrade": {
				"enum": [
					"Non-Technical Trekking Peak",
					"Technical Alpine Grade",
					"Extreme Technical Grade"
				],
				"type": "string"
			},
			"ExpeditionStatus": {
				"enum": [
					"active",
					"featured",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Expedition-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Expedition"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Expedition"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BlogStatus": {
				"enum": [
					"published",
					"draft",
					"archived"
				],
				"type": "string"
			},
...
			"ApiResponse_BlogArticle-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/BlogArticle"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/BlogArticle"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AssociateStatus": {
				"enum": [
					"active",
					"draft"
				],
				"type": "string"
			},
...
			"ApiResponse_Associate-Array_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/Associate"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/Associate"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Associate_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Associate"
							},
							{
								"items": {
									"$ref": "#/components/schemas/Associate"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAssociateDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"websiteUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/AssociateStatus"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAssociateDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"websiteUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/AssociateStatus"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_AdminLoginResponse_": {
				"properties": {
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/AdminLoginResponse"
							},
							{
								"items": {
									"$ref": "#/components/schemas/AdminLoginResponse"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminAuthSchema": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"data": {
						"items": {
							"type": "number",
							"enum": [
								null
							],
							"nullable": true
						},
						"type": "array",
						"nullable": true
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					},
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-nodejs-starter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/treks/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Trek-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL trekking packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Trekking"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TrekStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/tours/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tour-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL tour packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Tours"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tourType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TourType"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TourStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxDuration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/expeditions/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Expedition-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL expedition packages (including DRAFT, ACTIVE, and FEATURED).",
				"tags": [
					"Expeditions"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TripDifficulty"
						}
					},
					{
						"in": "query",
						"name": "climbingGrade",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ClimbingGrade"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ExpeditionStatus"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxAltitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPeakHeight",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/blogs/admin": {
			"get": {
				"operationId": "GetAdminAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BlogArticle-Array_"
								}
							}
						}
					}
				},
				"description": "Admin-only listing of ALL blog articles (including DRAFT, PUBLISHED, and ARCHIVED).",
				"tags": [
					"Blog Articles"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BlogStatus"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/associates": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Associate-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Associates & Partners"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/AssociateStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Associate_"
								}
							}
						}
					}
				},
				"tags": [
					"Associates & Partners"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAssociateDto"
							}
						}
					}
				}
			}
		},
		"/associates/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Associate_"
								}
							}
						}
					}
				},
				"tags": [
					"Associates & Partners"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Associate_"
								}
							}
						}
					}
				},
				"tags": [
					"Associates & Partners"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAssociateDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_boolean_"
								}
							}
						}
					}
				},
				"tags": [
					"Associates & Partners"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AdminLoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminAuthSchema"
							}
						}
					}
				}
			}
		},
		"/admin/auth/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AdminLoginResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/admin/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Admin Auth System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "https://abc.com/api/v1",
			"description": "prouduction server"
		},
		{
			"url": "http://localhost:4000/api/v1",
			"description": "Development server"
		}
	]
}